requirements/local.txt
# Local requirements file.
# Place items here that should ONLY be installed in DEVELOPMENT environments.

# include the base requirements file
-r _base.txt

django-debug-toolbar==1.3.0
django-extensions==1.5.5

# For easy deployments
# http://docs.fabfile.org/en/latest/
# Fabric

# Dependent on the Pypi version it can be required by Sphinx
sphinx_rtd_theme==0.1.8

# For building documentation
# http://sphinx-doc.org/
Sphinx==1.3.1

# For running tests
coverage==4.0a5

# For easy fixtures and test creation
model_mommy==1.2.4
mock==1.0.1

# For jenkins integration
django-jenkins==0.17.0
pep8==1.6.2
pyflakes==0.8.1
flake8==2.4.1

# Selenium Tests
selenium==2.45.0
